expenses: change ExpenseType from Enum to HashMap
The current ExpenseType uses Enum to store the different types of expenses according to the place of expenditure.

However, this does not provide any detail on the items purchased. This is also not scalable since it assumes that the products and services sold are fixed.

The different types of items that can be bought are now stored in a HashMap. Each item or service is assigned a unique menu number. The ExpenseType class contains two HashMaps, one for mapping menu number to item name, one for mapping menu number to price. Since a HashMap is now used, there needs to be error catching for cases where the user enters invalid menu numbers. Hence, a new exception ItemNotFoundException is made to account for such cases. Since the default prices of products are now stored as well, there is a new Expense constructor that only takes in the menu number, which can be used if the user wants to charge the default price for a product.

The toString() method in Expenses was also changed to use StringBuilder instead of string concatenation, since string concatenation is relatively expensive as compared to using StringBuilder, especially if the string grows large.

Note: apologies for this big commit, a repo reset was done on my side.