expense: remove unnecessary error checking
The constructors of Expense call the constructor of ExpenseType. Currently, the constructors of both classes check that the given item is in the menu.

This results in redundant code as the checks at ExpenseType will always pass, since only Expense calls the constructor of ExpenseType.

The checks at the Expense constructors are thus removed.