expense: add XML adaptability for Expenses and Expense
The Expenses for a room will have to be stored on the hard disk, since that information has to be kept between sessions. However, the current Expenses and Expense classes are not suitable to be written as XML.

This commit incorporates all the necessary code to store Expenses in addressbook.xml. However, the Expenses have been tagged to the Guest as of now, while pending the approval of another PR. The code here will be edited again when the storage of Rooms on XML has been implemented.