expense: implement XML storage for menu
The menu was previously defined as a static HashMap inside the ExpenseType class.

However, the menu should ideally be stored in a human-editable format, such as in an XML file together with guests and rooms data.

The menu has now been implemented with the ability to be stored in XML format. A separate class for the menu has been created, thus ExpenseType will not contain the menu anymore. The  menu is essentially a mapping of item numbers to their respective ExpenseType objects. The Menu object will only be accessible from the AddressBook or ModelManager, hence it will now be the responsibility of commands to check if a given item is present in the menu, otherwise the menu will be passed through the parameters to the relevant classes. Previously, ExpenseType contains the menu and thus it is responsible for conducting these checks. The ExpenseType now does not know of the existence of the menu, and thus will not be involved in checking. The type field in Expense objects will reference the same set of ExpenseType objects. In the event where the XML file contains an Expense that contains an ExpenseType that is not in the given menu, a new ExpenseType object with "Unknown" name will be created, thus being the only exception where an Expense object does not reference an ExpenseType in the menu. Tests have not been implemented.