model: change exception thrown by XmlAdaptedExpenseType
Previously, the toModelType() method in XmlAdaptedExpenseType can throw an IllegalArgumentException if the XmlAdaptedExpenseType object has an empty string in one of its fields.

However, the MainApp does not catch toModelType() throwing an IllegalArgumentException, hence preventing the initialisation of the application.

The exception is now caught and an IllegalValueException is thrown instead, which can be caught.