expenses: add Money class
Currently, money is represented in Expense and ExpenseType using double values.

However, in real life, money usually is of a certain format, such as always having two decimal places.

A new Money class has been added in order to capture this requirement. Expense and ExpenseType have incorporated Money into their itemCost fields. In their respective XmlAdapted representations, the cost is now stored as a String to preserve the decimal places of the cost. Tests for the existing classes have been adopted to the new Money class. More tests still need to be written for Money itself.