expenses: add more Money validation checks and tests
New validation checks were added for Money, namely that the dollar part of Money should be restricted to the range of Integer.MAX_VALUE and that the string representation should not have unnecessary leading zeroes for the dollar part (i.e. "01234.56" is not ok, "0.45" is ok). Tests for Money have also been added.