expenses: add support for negative money
Previously, it was assumed that monetary values will always be positive.

However, this may not be the case since negative monetary amounts may be required, for example, vouchers may be implemented as Menu items with a negative cost.

The Money class now accepts negative monetary values. Tests have also been added to account for this.