ServiceCommand: fix bug where new Expense items were not written
There was a bug where using ServiceCommand to add an Expense did not reflect in the GUI nor was it written into the XML file.

As such, the ServiceCommand basically did not do anything.

It was found that this was due to the code for adding expenses in Concierge.java and Room.java being outdated, as they did not employ functional programming. The bug has been fixed and a rough testing on the application shows that ServiceCommand is more or less functional. Tests will now have to be written. Also, minor text fixes were made, regarding the use of address when adding a guest.