XmlSerializableConcierge: change menu storage to list
Currently, the menu is being stored in the XML storage file in the form of a HashMap<String, XmlAdaptedExpenseType>, where the key is the item number and the value is the corresponding expense type.

There is data redundancy as the item number appears in both the key and value. In addition, undesirable things may happen if the item number in the key is different from the item number in the value.

Such coupling has been removed, by storing the menu as a list instead. Thus, the item number only appears once for every entry, and every item number can be assumed to be correct and independent of other item numbers. In addition, additional checks were added to ensure that the menu does not contain items with duplicate item numbers.