Room: fix addExpense bug
Currently, there is no exception checking for adding expenses at the room level. This means that rooms that are not checked in or have no bookings are able to have expenses added to it. This is because the only checks currently reside at the Concierge level. In addition, there are no checks in the XML conversion for cases where expenses appear at rooms that are not checked in or have no bookings.

The lack of checks will lead to mysterious behaviour, such as a room without guests having expenses (which is in fact what the current sample Concierge generates). Users may also be able to edit the XML files to give expenses to rooms with no guests.

These bugs have been fixed by moving the checks from Concierge.java's addExpense method to Room.java's addExpense method. For the XML conversion, additional checks have been added for XmlAdaptedRoom.java. The corresponding tests have also been updated.