Expenses: fix toStringTotalCost method throwing MoneyLimitExceededException
It was reported that entering a cost that is too high for service will cause the RoomDetailedPanel to disappear. This was discovered to be due to MoneyLimitExceededException, which was thrown in the method toStringTotalCost in Expenses, which naively used a Money object to store the total cost, even though it is possible for the total cost to exceed the limit of values that a Money object can store. The method has been fixed to use BigDecimal to avoid this problem, assuming that BigDecimal can have virtually unlimited length.