XmlAdaptedRoom: fix IllegalValueException due to expenses present in a checked-in but expired booking
The current implementation of XmlAdaptedRoom will only allow expenses to be added to rooms that have active and checked-in bookings.

However, this does not reflect the real-life scenario. For example, the user might exit the app on the same day as the last day of a booking with expenses, without checking out the room. The next day, the booking will become expired, and starting the app will thus throw an error. This behaviour should be prevented as it is possible that users may forget to check out after the day.

The section on adding expenses has been adjusted to allow for this. Expenses can now be added as long as there is a booking, and that booking is checked-in.