SampleDataUtil: add two more expense types into sample
Two expense types were added for the purpose of representing discounts and typo corrections. This is necessary since there is no command for editing or deleting an expense, hence any adjustment will have to be added as an expense with either positive or negative value. These two expense types are thus general expense types for such a purpose.