UG and DG updates (#252)
* XmlAdaptedRoom: fix IllegalValueException due to expenses present in a checked-in but expired booking

The current implementation of XmlAdaptedRoom will only allow expenses to be added to rooms that have active and checked-in bookings.

However, this does not reflect the real-life scenario. For example, the user might exit the app on the same day as the last day of a booking with expenses, without checking out the room. The next day, the booking will become expired, and starting the app will thus throw an error. This behaviour should be prevented as it is possible that users may forget to check out after the day.

The section on adding expenses has been adjusted to allow for this. Expenses can now be added as long as there is a booking, and that booking is checked-in.

* SampleDataUtil: add two more expense types into sample

Two expense types were added for the purpose of representing discounts and typo corrections. This is necessary since there is no command for editing or deleting an expense, hence any adjustment will have to be added as an expense with either positive or negative value. These two expense types are thus general expense types for such a purpose.

* UserGuide: update service section with the two adjustment expense types

A description for the usage of the two adjustment expense types has been added.

* UserGuide: fix minor stuff

* DeveloperGuide: add ServiceCommand stuff with a flowchart

Stuff was added. May add more stuff.

* ModelClassDiagram: update with current structure

The ModelClassDiagram illustrated the Address Book structure. The diagram has been updated to reflect our Concierge structure.

* ModelClassDiagram: update the better diagram also

The other diagram was updated with our current terminology. UniqueRoomList and Menu were not included, otherwise it would be too cluttered (and also because I'm lazy).

* DeveloperGuide: add sequence diagram for addExpense

Sequence diagram for the entire process of adding an expense is added. This is to supplement the flowchart with more detail on how an expense is added.

* DeveloperGuide: make addExpense sequence diagram bigger

The words are a bit unreadable.

* DeveloperGuide: add more design considerations

* all: fix checkstyle errors

* docs: add my own PPP

* XmlAdaptedRoomTest: add more tests to increase coverage

Tests targeted at some of the highlighted parts in Coveralls. Not sure why the other highlighted parts were not covered.

* XmlAdaptedRoomTest: fix checkstyle errors

* UserGuide: add minor corrections

* UserGuide and DeveloperGuide: minor edits from previous PR

* UserGuide: use important icon

* UserGuide: fix some formatting errors